requirements.txt
fastapi==0.81.0
uvicorn==0.18.3
marshmallow-sqlalchemy==0.29.0
psycopg2-binary==2.9.7
pytest
pytest-mock